chore(popup): update cypress tests
