+++ packages/popup/cypress/support/cmd/index.ts
@@ -1,4 +1,3 @@
 import "./browser";
-import "./login";
 import "./paste";
 import "./visit";
