+++ packages/popup/cypress/support/selectors.ts
@@ -1,33 +1,4 @@
 export const selectors = {
-  account: {
-    container: () => cy.findByTestId("account-container"),
-    dangerZone: {
-      deleteAccountButton: () => cy.findByTestId("delete-account-button"),
-      deleteAccountConfirmationAlert: () =>
-        cy.findByTestId("delete-account-confirmation-alert"),
-      deleteAccountConfirmationCancelButton: () =>
-        cy.findByTestId("delete-account-confirmation-cancel-button"),
-      deleteAccountConfirmationCheckbox: () =>
-        cy.findByTestId("delete-account-confirmation-checkbox"),
-      deleteAccountConfirmationContainer: () =>
-        cy.findByTestId("delete-account-confirmation-container"),
-      deleteAccountContainer: () => cy.findByTestId("delete-account-container"),
-      deleteAllRulesButton: () => cy.findByTestId("delete-all-rules-button"),
-      deleteAllRulesConfirmationAlert: () =>
-        cy.findByTestId("delete-all-rules-confirmation-alert"),
-      deleteAllRulesConfirmationCancelButton: () =>
-        cy.findByTestId("delete-all-rules-confirmation-cancel-button"),
-      deleteAllRulesConfirmationCheckbox: () =>
-        cy.findByTestId("delete-all-rules-confirmation-checkbox"),
-      deleteAllRulesConfirmationContainer: () =>
-        cy.findByTestId("delete-all-rules-confirmation-container"),
-      deleteAllRulesConfirmationSubmitButton: () =>
-        cy.findByTestId("delete-all-rules-confirmation-submit-button"),
-      deleteAllRulesContainer: () =>
-        cy.findByTestId("delete-all-rules-container"),
-      root: () => cy.findByTestId("danger-zone"),
-    },
-  },
   addNewRuleButton: () => cy.findByTestId("add-new-rule-button"),
   colorSelect: {
     dropdownButton: () => cy.findByTestId("color-select-dropdown-button"),
@@ -87,6 +58,32 @@ export const selectors = {
     },
   },
   options: {
+    dangerZone: {
+      deleteAccountButton: () => cy.findByTestId("delete-account-button"),
+      deleteAccountConfirmationAlert: () =>
+        cy.findByTestId("delete-account-confirmation-alert"),
+      deleteAccountConfirmationCancelButton: () =>
+        cy.findByTestId("delete-account-confirmation-cancel-button"),
+      deleteAccountConfirmationCheckbox: () =>
+        cy.findByTestId("delete-account-confirmation-checkbox"),
+      deleteAccountConfirmationContainer: () =>
+        cy.findByTestId("delete-account-confirmation-container"),
+      deleteAccountContainer: () => cy.findByTestId("delete-account-container"),
+      deleteAllRulesButton: () => cy.findByTestId("delete-all-rules-button"),
+      deleteAllRulesConfirmationAlert: () =>
+        cy.findByTestId("delete-all-rules-confirmation-alert"),
+      deleteAllRulesConfirmationCancelButton: () =>
+        cy.findByTestId("delete-all-rules-confirmation-cancel-button"),
+      deleteAllRulesConfirmationCheckbox: () =>
+        cy.findByTestId("delete-all-rules-confirmation-checkbox"),
+      deleteAllRulesConfirmationContainer: () =>
+        cy.findByTestId("delete-all-rules-confirmation-container"),
+      deleteAllRulesConfirmationSubmitButton: () =>
+        cy.findByTestId("delete-all-rules-confirmation-submit-button"),
+      deleteAllRulesContainer: () =>
+        cy.findByTestId("delete-all-rules-container"),
+      root: () => cy.findByTestId("danger-zone"),
+    },
     renderRate: {
       inputWrapper: () => cy.findByTestId("render-rate-input-wrapper"),
       toggleButton: () => cy.findByTestId("render-rate-toggle-button"),
